requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.13.5
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comtypes==1.4.12
ddgs==9.6.0
distro==1.9.0
edge-tts==7.2.3
frozenlist==1.7.0
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google_search_results==2.4.2
googleapis-common-protos==1.70.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.11.0
lxml==6.0.1
MouseInfo==0.1.3
multidict==6.6.4
primp==0.15.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.11.9
pydantic_core==2.33.2
pygame==2.6.1
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyparsing==3.2.4
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.1.1
pyttsx3==2.99
pytweening==1.2.0
pywin32==311
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
SpeechRecognition==3.14.3
tabulate==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
webrtcvad-wheels @ file://./wheels/webrtcvad_wheels-2.0.14-cp312-cp312-win_amd64.whl

wikipedia==1.4.0
yarl==1.20.1
